refactor ShouldDisplaySurveyPrompt from config package to survey
