+++ pkg/skaffold/config/util.go
@@ -307,12 +307,7 @@ func IsUpdateCheckEnabled(configfile string) bool {
 	return cfg == nil || cfg.UpdateCheck == nil || *cfg.UpdateCheck
 }
 
-func ShouldDisplaySurveyPrompt(configfile string) bool {
-	cfg, disabled := isSurveyPromptDisabled(configfile)
-	return !disabled && !recentlyPromptedOrTaken(cfg)
-}
-
-func isSurveyPromptDisabled(configfile string) (*ContextConfig, bool) {
+func IsSurveyPromptDisabled(configfile string) (*ContextConfig, bool) {
 	cfg, err := GetConfigForCurrentKubectx(configfile)
 	if err != nil {
 		return nil, false
@@ -320,11 +315,11 @@ func isSurveyPromptDisabled(configfile string) (*ContextConfig, bool) {
 	return cfg, cfg != nil && cfg.Survey != nil && cfg.Survey.DisablePrompt != nil && *cfg.Survey.DisablePrompt
 }
 
-func recentlyPromptedOrTaken(cfg *ContextConfig) bool {
+func RecentlyPromptedOrTaken(cfg *ContextConfig, current time.Time) bool {
 	if cfg == nil || cfg.Survey == nil {
 		return false
 	}
-	return lessThan(cfg.Survey.LastTaken, 90*24*time.Hour) || lessThan(cfg.Survey.LastPrompted, 10*24*time.Hour)
+	return lessThanWithTime(cfg.Survey.LastTaken, 90*24*time.Hour, current) || lessThanWithTime(cfg.Survey.LastPrompted, 10*24*time.Hour, current)
 }
 
 func ShouldDisplayUpdateMsg(configfile string) bool {
@@ -363,13 +358,17 @@ func UpdateMsgDisplayed(configFile string) error {
 	return err
 }
 
-func lessThan(date string, duration time.Duration) bool {
+func lessThanWithTime(date string, duration time.Duration, current time.Time) bool {
 	t, err := time.Parse(time.RFC3339, date)
 	if err != nil {
 		logrus.Debugf("could not parse date %q", date)
 		return false
 	}
-	return current().Sub(t) < duration
+	return current.Sub(t) < duration
+}
+
+func lessThan(date string, duration time.Duration) bool {
+	return lessThanWithTime(date, duration, current())
 }
 
 func UpdateGlobalSurveyTaken(configFile string) error {
